// api/routes/horses/models/HorsesResponse.ts
export interface HorseResponse {
    horseNumberUk: string;
    horseNumber: number;
    lifeStatus?: string;
    deathEvent?: string;
    deathDate?: number;
    birthYear?: number;
    birthDate?: string;
    age?: number;
    birthCountry?: string;
    birthCountryLib?: string;
    birthName: string;
    coatColorLib?: string;
    coatColorCode?: string;
    sexCode?: "M" | "F" | "H";
    breedCode?: string;
    breedLib?: string;
    morphology?: string;
    nusirecle?: string;
    ueln?: string;
    raceFamily?: string;
    nature?: string;
    fictif?: string;
    checkStatus?: string;
    motherNumber?: number;
    motherBreed?: string;
    motherCountry?: string;
    motherName?: string;
    fatherNumber?: number;
    fatherBreed?: string;
    fatherCountry?: string;
    fatherName?: string;
    pdmNumber?: number;
    pdmBreed?: string;
    pdmName?: string;
    personalNumberBirth?: string;
    birthParts?: string;
    birthOrder?: string;
    docWeb?: string;
    nuraces3?: string;
    coraces3?: string;
    nuraces3P?: string;
    coraces3P?: string;
    nuraces3M?: string;
    coraces3M?: string;
    breeders?: string[];
}